pI.py: add addCurrentViewpoint()
This function shall read the current 3D View (View3DInventor) and compile the selected items, the color of the selection etc. into a viewpoint object. Currently the function reads the camera's position and orientation and creates a orthogonal camera or perspective camera object respectively, which is then added to a new viewpoint object. that is the only member that will be set atm.

+++ bcfplugin/rdwr/viewpoint.py
@@ -177,7 +177,7 @@ class PerspectiveCamera(Camera, XMLName):
             viewPoint: Point,
             direction: Direction,
             upVector: Direction,
-            fieldOfView: int,
+            fieldOfView: float,
             containingElement = None,
             state: State.States = State.States.ORIGINAL):
         Camera.__init__(self,
